+++ lib/Datepicker/stories/DatepickerDemo.js
@@ -18,17 +18,40 @@ class DatepickerDemo extends React.Component {
     }
 
     this.setState({
-      [name]: nextDate,
+      // [name]: nextDate,
+      [name]: value
     });
+  }
+
+  rfUpdate = (value) => {
+    this.setState({
+      rfd1: nextDate,
+    });
+  }
+
+  clearDate = () => {
+    const newState = {};
+    for(const p in this.state) {
+      newState[p] = '';
+    }
+    this.setState(newState);
   }
 
   render() {
     return (
       <div>
+        <button onClick={this.clearDate} type="button">Clear dates</button>
         <Datepicker
           label="Date"
           value={this.state.day1}
           onChange={(e) => { this.onChangeDate(e.target.value, 'day1'); }}
+        />
+        <Datepicker
+          label="RFF-style-Date"
+          input={{
+            value: this.state.day1,
+            onChange: this.rfUpdate
+          }}
         />
         <br />
         <Datepicker
@@ -37,7 +60,7 @@ class DatepickerDemo extends React.Component {
           onChange={(e) => { this.onChangeDate(e.target.value, 'day3'); }}
           readOnly
         />
-        <br />
+        {/*  <br />
         <Datepicker
           label="Date (disabled)"
           disabled
@@ -48,7 +71,7 @@ class DatepickerDemo extends React.Component {
           value={this.state.day2}
           onChange={(e) => { this.onChangeDate(e.target.value, 'day2'); }}
           required
-        />
+        /> */}
       </div>
     );
   }
